refactor: signInSlice -> userSlice
